@@ -32,7 +32,7 @@
     "colors": "^1.1.2",
     "cross-env": "^2.0.1",
     "css-loader": "^0.25.0",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^2.0.0-rc.3",
     "gulp": "^3.9.0",
     "gulp-autoprefixer": "^3.1.0",
     "gulp-babel": "^6.1.2",
@@ -52,6 +52,7 @@
     "morgan": "^1.7.0",
     "node-sass": "^4.1.1",
     "postcss-loader": "^0.13.0",
+    "react-hot-loader": "^3.0.0-beta.6",
     "redux-devtools": "^3.3.1",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.0.11",
@@ -68,14 +69,14 @@
     "vue": "^2.1.10",
     "vue-hot-reload-api": "^2.0.8",
     "vue-html-loader": "^1.2.3",
-    "vue-loader": "^10.0.2",
+    "vue-loader": "^10.3.0",
     "vue-router": "^2.1.3",
     "vue-server-renderer": "^2.1.10",
     "vue-style-loader": "^1.0.0",
     "vue-template-compiler": "^2.1.10",
     "vueify": "^8.7.0",
     "vuex": "^2.1.1",
-    "webpack": "^1.13.1",
+    "webpack": "^2.2.1",
     "webpack-dev-middleware": "^1.9.0",
     "webpack-dev-server": "^1.16.2",
     "webpack-hot-middleware": "^2.15.0",
@@ -90,8 +91,8 @@
     "fs": "0.0.2",
     "lowdb": "^0.14.0",
     "mysql": "^2.10.2",
-    "react": "^0.14.8",
-    "react-dom": "^0.14.8",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
     "react-redux": "^4.4.5",
     "react-router": "^2.2.4",
     "redux": "^3.5.2"
