@@ -1,10 +1,12 @@
+import dir from './config.js';
+
 export default [{
-	background: DIR.img + '/index/banner_0.jpg',
+	background: dir.img + '/index/banner_0.jpg',
 	href: '#/shopping_cart'
 },{
-	background: DIR.img + '/index/banner_1.jpg'
+	background: dir.img + '/index/banner_1.jpg'
 },{
-	background: DIR.img + '/index/banner_2.jpg'
+	background: dir.img + '/index/banner_2.jpg'
 },{
-	background: DIR.img + '/index/banner_3.jpg'
+	background: dir.img + '/index/banner_3.jpg'
 }];
