@@ -12,7 +12,7 @@ var applyMiddleware = Redux.applyMiddleware;
 var compose = Redux.compose;
 var combineReducers =  Redux.combineReducers;
 
-var $$reducer = Redux.combineReducers({
+var rootReducer = Redux.combineReducers({
 	user,
 	shoppingCart,
 	notice,
@@ -22,17 +22,12 @@ var $$reducer = Redux.combineReducers({
 	item
 });
 
-// actions
-var LOGIN = {type:'LOGIN'};
-var INCREMENT = {type:'INCREMENT'};
-var DECREMENT = {type:'DECREMENT'};
-
 // store
-const $$store = createStore(
-	$$reducer,
+const store = createStore(
+	rootReducer,
 	applyMiddleware(ReduxThunk.default)
 	/*,enhancer*/
 );
-console.debug( 'Redux: store created.',$$store.getState() )
+console.debug( 'Redux: store created.',store.getState() )
 
-export {$$store};
+export default store;
